+++ cmd/list.go
@@ -11,11 +11,11 @@ import (
 	"text/tabwriter"
 	"time"
 
+	klausoci "github.com/giantswarm/klaus-oci"
 	"github.com/spf13/cobra"
 
 	"github.com/giantswarm/klausctl/pkg/config"
 	"github.com/giantswarm/klausctl/pkg/instance"
-	"github.com/giantswarm/klausctl/pkg/oci"
 	"github.com/giantswarm/klausctl/pkg/runtime"
 )
 
@@ -163,7 +163,7 @@ func toolchainLabel(cfg *config.Config) string {
 }
 
 func shortToolchain(image string) string {
-	repo := oci.RepositoryFromRef(image)
+	repo := klausoci.RepositoryFromRef(image)
 	name := filepath.Base(repo)
 	if strings.HasPrefix(name, "klaus-") {
 		return strings.TrimPrefix(name, "klaus-")
@@ -175,7 +175,7 @@ func shortRefName(ref string) string {
 	if ref == "" {
 		return ""
 	}
-	return filepath.Base(oci.RepositoryFromRef(ref))
+	return filepath.Base(klausoci.RepositoryFromRef(ref))
 }
 
 func valueOrDash(v string) string {